Refactor and update mobile responsive
