+++ src/assets/css/careers/careersPage.css
@@ -17,16 +17,36 @@
     align-self: flex-start; 
     flex-direction: column;
     
-    flex-basis: 367px;
-    flex-grow: 0;
-    flex-shrink: 0;
-    
     row-gap: 10px; 
     align-items: center; 
     background-color: #FFFFFF; 
     border-radius: 20px; 
     padding: 50px;
 }
+
+.career_card_size{
+    flex-basis: 367px;
+    flex-grow: 0;
+    flex-shrink: 0;
+}
+
+@media screen and (max-width: 390px){
+    .career_card_size{
+        flex-basis: 320px;
+        flex-grow: 0;
+        flex-shrink: 0;
+    }
+}
+
+@media screen and (max-width: 320px) {
+    .career_card_size{
+        flex-basis: 270px;
+        flex-grow: 0;
+        flex-shrink: 0;
+    }
+}
+
+
 
 .career_logo{
     height: 35px; width: auto; margin-top: 10px
